v0.32.1

See merge request shaninalex/lumna!80

+++ app/api/middlewares/auth.go
@@ -50,7 +50,7 @@ func AuthMiddleware(c *gin.Context) {
 	}
 
 	ctx := c.Request.Context()
-	ctx = context.WithValue(ctx, internal.ContextIdentity, userID)
+	ctx = context.WithValue(ctx, internal.ContextIdentity, uint(userID))
 	// fill request context with values here
 
 	c.Set("userID", uint(userID))
